# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: WorkTime Sync API
  version: 0.1.0
  description: |
    Веб-платформа для актуализации рабочего времени сотрудников.
    Endpoints разделены на публичные (`/auth/*`, `/webhooks/*`) и защищённые JWT (всё остальное).
  contact:
    name: WorkTime Sync team
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /api/v1/integrations:
    get:
      tags: [integrations]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: ok }
  /api/v1/integrations/ical:
    post:
      tags: [integrations]
      security: [{ bearerAuth: [] }]
      summary: Подключить iCal/ICS feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                feed_url: { type: string }
                label: { type: string }
      responses:
        '201': { description: created }
  /api/v1/integrations/caldav:
    post:
      tags: [integrations]
      security: [{ bearerAuth: [] }]
      summary: Подключить CalDAV (Yandex/Apple/Nextcloud)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, username, password]
              properties:
                endpoint: { type: string }
                username: { type: string }
                password: { type: string }
                cal_path: { type: string }
                label: { type: string }
      responses:
        '201': { description: created }
...
  /api/v1/recommendations:
    get:
      tags: [recommendations]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: ok }
  /api/v1/recommendations/generate:
    post:
      tags: [recommendations]
      security: [{ bearerAuth: [] }]
      summary: Пересобрать рекомендации (rule-based + GigaChat)
      responses:
        '200': { description: ok }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...